Move clean_simulation_directory to run_four_c
